requirements.txt
pydantic==2.7.4
PyYAML==6.0.2
Pillow==10.4.0
requests==2.32.4
tzdata
pygame
colorama